Rename components folder
