+++ src/app/app.module.ts
@@ -7,7 +7,7 @@ import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MaterialModule } from '@angular/material';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { GlobalHeaderModule } from './componnents/global-header/header.module';
+import { GlobalHeaderModule } from './components/global-header/header.module';
 import { VacanciesDetailModule } from './vacancy-detail/vacancy-detail-wrapper/vacancy-detail-wrapper.module';
 import 'hammerjs';
