+++ src/app/vacancy-detail/vacancy-detail-content/vacancy-detail-content.module.ts
@@ -2,9 +2,9 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { VacancyDetailContentComponent } from './vacancy-detail-content.component';
-import { InputTextModule } from '../../componnents/input-text/input-text.module';
-import { DatepickerModule } from '../../componnents/datepicker/datepicker.module';
-import { InputTextAreaModule } from '../../componnents/input-text-area/input-text-area.module';
+import { InputTextModule } from '../../components/input-text/input-text.module';
+import { DatepickerModule } from '../../components/datepicker/datepicker.module';
+import { InputTextAreaModule } from '../../components/input-text-area/input-text-area.module';
 
 @NgModule({
   imports: [
